Refactor Shareable
