--- CrocoDash/shareable/apply.py
@@ -1,100 +1,0 @@
-from visualCaseGen.custom_widget_types.case_tools import xmlchange, append_user_nl
-import shutil
-from pathlib import Path
-from CrocoDash.forcing_configurations import *
-from CrocoDash.logging import setup_logger
-
-logger = setup_logger(__name__)
-
-INPUTDIR_FILE_PREFIXES = ("forcing_obc_segment_", "init_")
-
-
-def copy_xml_files_from_case(old_caseroot, new_caseroot, filenames):
-    old_caseroot = Path(old_caseroot)
-    new_caseroot = Path(new_caseroot)
-    for name in filenames:
-        logger.info(f"Copying {old_caseroot / name} into new caseroot")
-        shutil.copy(old_caseroot / name, new_caseroot / name)
-
-
-def copy_user_nl_params_from_case(
-    old_caseroot,
-    usernlparams,
-):
-    for key in usernlparams:
-        usernl = Path(old_caseroot) / f"user_nl_{key}"
-        with usernl.open() as f:
-            for line in f:
-                line = line.strip()
-                if line.startswith("!") or "=" not in line:
-                    continue
-
-                # PARAM=VALUE
-                param, value = line.split("=", 1)
-                param = param.split()[
-                    0
-                ]  # Get just the param name without any trailing comments
-                if param in usernlparams[key]:
-                    logger.info(f"Adding {param}={value} into user_nl_{key}")
-                    append_user_nl(key, [(param, value)], do_exec=True)
-
-
-def copy_source_mods_from_case(
-    old_caseroot,
-    new_caseroot,
-    short_filepaths,
-):
-    old_caseroot = Path(old_caseroot)
-    new_caseroot = Path(new_caseroot)
-    for path in short_filepaths:
-        path = Path(path)
-        logger.info(f"Adding {old_caseroot / 'SourceMods' / path} into new caseroot")
-        shutil.copy(
-            old_caseroot / "SourceMods" / path, new_caseroot / "SourceMods" / path
-        )
-
-
-def apply_xmlchanges_to_case(
-    old_caseroot,
-    xmlchangeparams,
-):
-    # Copy old_caseroot replya.sh with that line on it and replay it verbatim with xmlchange()
-    replay = Path(old_caseroot) / "replay.sh"
-    with replay.open() as f:
-        for line in f:
-            line = line.strip()
-
-            if not line.startswith("./xmlchange"):
-                continue
-
-            # ./xmlchange PARAM=VALUE
-            _, kv = line.split(None, 1)
-            param, value = kv.split("=", 1)
-            if param in xmlchangeparams:
-                logger.info(f"Running {param}={value} xmlchange to new caseroot")
-                xmlchange(param, value, is_non_local=True)
-
-
-def copy_configurations_to_case(old_forcing_config, case, inputdir_ocnice):
-    """
-    Copy forcing configurations from inputdir_ocnice to case.inputdir/ocnice.
-    """
-
-    case_ocnice = case.inputdir / "ocnice"
-
-    for src in inputdir_ocnice.iterdir():
-        if src.is_file() and src.name.startswith(INPUTDIR_FILE_PREFIXES):
-            logger.info(f"Copying {src} to new case")
-            shutil.copy(src, case_ocnice)
-
-    # Iterate through old_forcing_config outputs
-    for key in old_forcing_config:
-        if key == "basic" or key not in case.fcr.active_configurators.keys():
-            continue
-        configurator = ForcingConfigRegistry.get_configurator(old_forcing_config[key])
-        for filepath in configurator.get_output_filepaths(
-            inputdir_ocnice
-        ):  # Get output Filepaths works as if this were a case, but this is a bundle
-            path = inputdir_ocnice / Path(filepath).name  # Get just the name
-            logger.info(f"Copying {path} to new case")
-            shutil.copy(path, case_ocnice)
